+++ test/parserTest/testCaseParserTest.test.ts
@@ -7,10 +7,13 @@ import * as path from 'path';
 import {buildFileToSuite} from '../../src/parsers/testCaseFileParser';
 import {TestSuite} from '../../src/robotModels/TestSuite';
 import {Keyword} from '../../src/robotModels/Keyword';
+import {TestCase} from '../../src/robotModels/TestCase';
 import {LibraryMetaData, ResourceMetaData} from '../../src/robotModels/MetaData';
 import {getSuiteFromFileName} from './parserUtil';
 
 // note that this extension don't care about test case table and variable table in suite.
+// all these files used for testing shouldn't be modified or moved
+// or these tests will run failed
 suite("File Builder Test", () => {
     let resourceSettingValue : string = "../test_resource";
     let expectedResourceMetaData : ResourceMetaData[] = [new ResourceMetaData("Resource", "../test_resource")];
@@ -115,6 +118,29 @@ suite("File Builder Test", () => {
         let suite : TestSuite = getSuiteFromFileName(fileName);
 
         assertVarTableExisted(suite);
+    });
+
+    test("test TestCase local variable assignment", () => {
+        let fileName : string = "testLocalVariable.txt";
+        let suite : TestSuite = getSuiteFromFileName(fileName);
+
+        let caseNumber : number = 4;
+        assert.equal(suite.testCases.length, caseNumber);
+        assert.equal(suite.testCases[0].startLine, 1);
+        assert.equal(suite.testCases[0].endLine, 5);
+        assert.equal(suite.testCases[0].variables.size, 1);
+
+        assert.equal(suite.testCases[1].startLine, 5);
+        assert.equal(suite.testCases[1].endLine, 10);
+        assert.equal(suite.testCases[1].variables.size, 2);
+
+        assert.equal(suite.testCases[2].startLine, 11);
+        assert.equal(suite.testCases[2].endLine, 14);
+        assert.equal(suite.testCases[2].variables.size, 1);
+
+        assert.equal(suite.testCases[3].startLine, 14);
+        assert.equal(suite.testCases[3].endLine, 16);
+        assert.equal(suite.testCases[3].variables.size, 0);
     });
 
 });